Run the Shiny Application

This function runs the Shiny application for exploring vehicle prices from 2023 to 2025. It provides interactive features to filter the data by year, price range, make, transmission, and body type. Users can view the distribution of vehicle prices and download the filtered dataset as a CSV file.

Examples

shinyApp(ui = ui, server = server)
#> Error in shinyApp(ui = ui, server = server): could not find function "shinyApp"